fixed the image upload and added the booking and package routes and controllers as well as the review model routes and controllers

+++ src/utils/cloudinary.js
@@ -1,6 +1,6 @@
-import {v2 as cloudinary} from "cloudinary"
-import fs from "fs"
-
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
+import { ApiError } from "./ApiError.js";
 
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
@@ -9,23 +9,19 @@ cloudinary.config({
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null
-        //upload the file on cloudinary
-        const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        // file has been uploaded successfull
-        //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
-        return response;
+  try {
+    if (!localFilePath) return null;
 
-    } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        throw new ApiError(500, "Failed to upload file to Cloudinary");
-    }
-}
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto"
+    });
 
+    fs.unlinkSync(localFilePath); // remove local file
+    return response; // Return the full response object
+  } catch (error) {
+    if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
+    throw new ApiError(500, "Failed to upload file to Cloudinary");
+  }
+};
 
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
